Implement executeLargeBatch for ldbc connector

+++ module/ldbc-connector/shared/src/main/scala/ldbc/connector/exception/BatchUpdateException.scala
@@ -20,7 +20,7 @@ import org.typelevel.otel4s.Attribute
  */
 class BatchUpdateException(
   message:          String,
-  updateCounts:     List[Int],
+  updateCounts:     List[Long],
   sqlState:         Option[String] = None,
   vendorCode:       Option[Int]    = None,
   sql:              Option[String] = None,
